+++ www/home.js
@@ -14,7 +14,7 @@
    information rather than as an interruption. */
 function capBanner(){
   if(can('words')) return '';
-  var left=FREE_LIMIT-WORDS.length;
+  var left=wordCap()-WORDS.length;
   if(left>20 || left<0) return '';
   return '<button class="capwarn"' + DO('go', ["plans"]) + '>'+t('cap.warn', left)+
     '<span class="capgo">'+t('up.cta')+ICON_GO+'</span></button>';